pruebas unitarias página web

+++ administrador/app/Http/Controllers/CarruselController.php
@@ -18,13 +18,13 @@ class CarruselController extends Controller
     /*======================================================
     =            Agregar nuevo item al carrusel            =
     ======================================================*/
-    
+
     /**
      * Como el carrusel es un dato JSON dentro de la tabla de página, agregar un nuevo dato
      * se hace por medio de una actualización en el mismo campo.
      *
      */
-    
+
 
     public function update($id, Request $request) {
 
@@ -47,6 +47,8 @@ class CarruselController extends Controller
         $indice = array('indice' => $request->input("indice"));
         /*echo '<pre>'; print_r($indice['indice']); echo '</pre>';
         return;*/
+        $escenario = array('escenario' => $request->input("escenario"));
+
         $categoria_Nuevo = array('categoria' => $request->input("categoria-".($indice['indice']+1)));
         $titulo_Nuevo = array('titulo' => $request->input("titulo-".($indice['indice']+1)));
         $texto_Nuevo = array('texto' => $request->input("texto-".($indice['indice']+1)));
@@ -57,7 +59,7 @@ class CarruselController extends Controller
         $foto_Delante_Nuevo = array('foto_Delante' => $request->input("foto-delante-".($indice['indice']+1)));
         $fondo_Nuevo = array('fondo' => $request->file("fondo-".($indice['indice']+1)));
 
-        
+
 
         if (!empty($categoria_Nuevo["categoria"]) && !empty($titulo_Nuevo["titulo"]) && !empty($texto_Nuevo["texto"]) && !empty($fondo_Nuevo["fondo"])) {
             $indice['indice'] = $indice['indice'] + 1; /*se incrementa la variable indice para agregar el nuevo indice*/
@@ -65,7 +67,7 @@ class CarruselController extends Controller
             $indice['indice'] = $indice['indice'] + 1;
         }
 
-        
+
         for ($i=0; $i <= $indice['indice']; $i++) {
 
             if (($imagenExterna['imagenExterna']=="si") && ($i == $indice['indice'])) {
@@ -77,7 +79,7 @@ class CarruselController extends Controller
                 ${"titulo_".$i} = array('titulo_'.$i => $request->input("titulo-".$i));
                 ${"texto_".$i} = array('texto_'.$i => $request->input("texto-".$i));
             }
-            
+
             ${"boton_1_Color_".$i} = array('boton_1_color'.$i => $request->input("boton-1-color-".$i));
             ${"boton_1_Texto_".$i} = array('boton_1_texto'.$i => $request->input("boton-1-texto-".$i));
             ${"url_Boton_1_".$i} = array('url_Boton_1_'.$i => $request->input("url-boton-1-".$i));
@@ -95,7 +97,7 @@ class CarruselController extends Controller
 
         $carrusel = "[{";
 
-        for ($i=0; $i <= $indice['indice']; $i++) { 
+        for ($i=0; $i <= $indice['indice']; $i++) {
             ${"validar_Categoria_".$i} = \Validator::make(${"categoria_".$i}, [ "categoria_".$i => 'required|regex:/^[,\\0-9a-zA-ZñÑáéíóúÁÉÍÓÚ ]+$/i']);
             ${"validar_Titulo_".$i} = \Validator::make(${"titulo_".$i}, ["titulo_".$i => 'required|regex:/^[=\\&\\$\\;\\-\\_\\*\\"\\<\\>\\?\\¿\\!\\¡\\:\\,\\.\\0-9a-zA-ZñÑáéíóúÁÉÍÓÚ ]+$/i']);
             ${"validar_Texto_".$i} = \Validator::make(${"texto_".$i}, ["texto_".$i => 'required|regex:/^[=\\&\\$\\;\\-\\_\\*\\"\\<\\>\\?\\¿\\!\\¡\\:\\,\\.\\0-9a-zA-ZñÑáéíóúÁÉÍÓÚ ]+$/i']);
@@ -104,7 +106,7 @@ class CarruselController extends Controller
                 return redirect("/pagina_web/ingresarCarrusel")->with("no-validacion", "");
             }
 
-            
+
             if (!empty(${"boton_1_Color_".$i}["boton_1_color".$i])) {
                 ${"validar_Boton_1_Color_".$i} = \Validator::make(${"boton_1_Color_".$i}, [ "boton_1_color".$i => 'required|regex:/^[-\\a-zA-Z]+$/i']);
                 if (${"validar_Boton_1_Color_".$i}->fails()) {
@@ -139,7 +141,7 @@ class CarruselController extends Controller
                 ${"validar_Url_Boton_2_Actual_".$i} = \Validator::make(${"url_Boton_2_".$i}, [ "url_Boton_2_".$i => 'required|regex:/^[=\\&\\$\\-\\_\\?\\!\\:\\.\\0-9a-zA-Z]+$/i']);
                 if (${"validar_Url_Boton_2_Actual_".$i}->fails()) {
                     return redirect("/pagina_web/ingresarCarrusel")->with("no-validacion", "");
-                }      
+                }
             }
             if (!empty(${"foto_Delante_Actual".$i}["foto_Delante_".$i])) {
                 ${"validar_Foto_Delante_Actual_".$i} = \Validator::make(${"foto_Delante_Actual_".$i}, [ "foto_Delante_".$i => 'required|regex:/^[=\\&\\$\\-\\_\\?\\!\\:\\.\\0-9a-zA-Z]+$/i']);
@@ -153,43 +155,52 @@ class CarruselController extends Controller
                     return redirect("/pagina_web/ingresarCarrusel")->with("no-validacion", "");
                 }
             }
-            
+
             if (!empty(${"foto_Delante_Temporal_".$i}["foto_Delante_".$i])) {
                 ${"validar_Foto_Delante_Temporal_".$i} = \Validator::make(${"foto_Delante_Temporal_".$i}, [ "foto_Delante_".$i => 'required|image|mimes:jpg,jpeg,png|max:2000000']);
                 if (${"validar_Foto_Delante_Temporal_".$i}->fails()) {
                     return redirect("/pagina_web/ingresarCarrusel")->with("no-validacion", "");
                 } else {
-                    if (!empty(${"foto_Delante_Actual_".$i}["foto_Delante_".$i])) {
-                        unlink(${"foto_Delante_Actual_".$i}["foto_Delante_".$i]);
-                    }
+
 
                     $aleatorio = mt_rand(10000, 99999);
                     ${"ruta_Foto_Delante_".$i} = "vistas/images/pagina_web/carrusel/".$aleatorio.".".${"foto_Delante_Temporal_".$i}["foto_Delante_".$i]->guessExtension();
 
-                    /*----------  Redimensionar imagen  ----------*/
-                    list($ancho, $alto) = getimagesize(${"foto_Delante_Temporal_".$i}["foto_Delante_".$i]);
-                    $nuevoAncho = 600;
-                    $nuevoAlto = 500;
-
-                    if((${"foto_Delante_Temporal_".$i}["foto_Delante_".$i]->guessExtension() == "jpeg") || (${"foto_Delante_Temporal_".$i}["foto_Delante_".$i]->guessExtension() == "jpg")){
-
-                        $origen = imagecreatefromjpeg(${"".$i}["".$i]);
-                        $destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
-                        imagecopyresized($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
-                        imagejpeg($destino, ${"ruta_Foto_Delante_".$i});
-
+                    if ($escenario["escenario"] == "prueba") {
+                        move_uploaded_file(${"foto_Delante_Temporal_".$i}["foto_Delante_".$i], ${"ruta_Foto_Delante_".$i});
                     }
 
-                    if(${"foto_Delante_Temporal_".$i}["foto_Delante_".$i]->guessExtension() == "png"){
+                    if ($escenario["escenario"] == "sistema") {
+                        if (!empty(${"foto_Delante_Actual_".$i}["foto_Delante_".$i])) {
+                            unlink(${"foto_Delante_Actual_".$i}["foto_Delante_".$i]);
+                        }
 
-                        $origen = imagecreatefrompng(${"foto_Delante_Temporal_".$i}["foto_Delante_".$i]);
-                        $destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
-                        imagealphablending($destino, FALSE); 
-                        imagesavealpha($destino, TRUE);
-                        imagecopyresampled($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
-                        imagepng($destino, ${"ruta_Foto_Delante_".$i});
-                        
+                        /*----------  Redimensionar imagen  ----------*/
+                        list($ancho, $alto) = getimagesize(${"foto_Delante_Temporal_".$i}["foto_Delante_".$i]);
+                        $nuevoAncho = 600;
+                        $nuevoAlto = 500;
+
+                        if((${"foto_Delante_Temporal_".$i}["foto_Delante_".$i]->guessExtension() == "jpeg") || (${"foto_Delante_Temporal_".$i}["foto_Delante_".$i]->guessExtension() == "jpg")){
+
+                            $origen = imagecreatefromjpeg(${"".$i}["".$i]);
+                            $destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
+                            imagecopyresized($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
+                            imagejpeg($destino, ${"ruta_Foto_Delante_".$i});
+
+                        }
+
+                        if(${"foto_Delante_Temporal_".$i}["foto_Delante_".$i]->guessExtension() == "png"){
+
+                            $origen = imagecreatefrompng(${"foto_Delante_Temporal_".$i}["foto_Delante_".$i]);
+                            $destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
+                            imagealphablending($destino, FALSE);
+                            imagesavealpha($destino, TRUE);
+                            imagecopyresampled($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
+                            imagepng($destino, ${"ruta_Foto_Delante_".$i});
+
+                        }
                     }
+
                 }
             } else {
                 ${"ruta_Foto_Delante_".$i} = ${"foto_Delante_Actual_".$i}["foto_Delante_".$i];
@@ -199,37 +210,45 @@ class CarruselController extends Controller
                 if (${"validar_Fondo_Temporal_".$i}->fails()) {
                     return redirect("/pagina_web/ingresarCarrusel")->with("no-validacion", "");
                 } else {
-                    if (!empty(${"fondo_Actual_".$i}["fondo_".$i])) {
-                        unlink(${"fondo_Actual_".$i}["fondo_".$i]);
-                    }
 
                     $aleatorio = mt_rand(10000, 99999);
                     ${"ruta_Fondo_".$i} = "vistas/images/pagina_web/carrusel/".$aleatorio.".".${"fondo_Temporal_".$i}["fondo_".$i]->guessExtension();
 
-                    /*----------  Redimensionar imagen  ----------*/
-                    list($ancho, $alto) = getimagesize(${"fondo_Temporal_".$i}["fondo_".$i]);
-                    $nuevoAncho = 2000;
-                    $nuevoAlto = 1333;
-
-                    if((${"fondo_Temporal_".$i}["fondo_".$i]->guessExtension() == "jpeg") || (${"fondo_Temporal_".$i}["fondo_".$i]->guessExtension() == "jpg")){
-
-                        $origen = imagecreatefromjpeg(${"fondo_Temporal_".$i}["fondo_".$i]);
-                        $destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
-                        imagecopyresized($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
-                        imagejpeg($destino, ${"ruta_Fondo_".$i});
-
+                    if ($escenario["escenario"] == "prueba") {
+                        move_uploaded_file(${"fondo_Temporal_".$i}["fondo_".$i], ${"ruta_Fondo_".$i});
                     }
 
-                    if(${"fondo_Temporal_".$i}["fondo_".$i]->guessExtension() == "png"){
+                    if ($escenario["escenario"] == "sistema") {
+                        if (!empty(${"fondo_Actual_".$i}["fondo_".$i])) {
+                            unlink(${"fondo_Actual_".$i}["fondo_".$i]);
+                        }
 
-                        $origen = imagecreatefrompng(${"fondo_Temporal_".$i}["fondo_".$i]);
-                        $destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
-                        imagealphablending($destino, FALSE); 
-                        imagesavealpha($destino, TRUE);
-                        imagecopyresampled($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
-                        imagepng($destino, ${"ruta_Fondo_".$i});
-                        
+                        /*----------  Redimensionar imagen  ----------*/
+                        list($ancho, $alto) = getimagesize(${"fondo_Temporal_".$i}["fondo_".$i]);
+                        $nuevoAncho = 2000;
+                        $nuevoAlto = 1333;
+
+                        if((${"fondo_Temporal_".$i}["fondo_".$i]->guessExtension() == "jpeg") || (${"fondo_Temporal_".$i}["fondo_".$i]->guessExtension() == "jpg")){
+
+                            $origen = imagecreatefromjpeg(${"fondo_Temporal_".$i}["fondo_".$i]);
+                            $destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
+                            imagecopyresized($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
+                            imagejpeg($destino, ${"ruta_Fondo_".$i});
+
+                        }
+
+                        if(${"fondo_Temporal_".$i}["fondo_".$i]->guessExtension() == "png"){
+
+                            $origen = imagecreatefrompng(${"fondo_Temporal_".$i}["fondo_".$i]);
+                            $destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
+                            imagealphablending($destino, FALSE);
+                            imagesavealpha($destino, TRUE);
+                            imagecopyresampled($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
+                            imagepng($destino, ${"ruta_Fondo_".$i});
+
+                        }
                     }
+
                 }
             } else {
                 ${"ruta_Fondo_".$i} = ${"fondo_Actual_".$i}["fondo_".$i];
@@ -241,7 +260,7 @@ class CarruselController extends Controller
                 $carrusel = $carrusel."\n\t".'"categoria": "'.${'categoria_'.$i}['categoria_'.$i].'",'."\n\t".'"titulo": "'.${'titulo_'.$i}['titulo_'.$i].'",'."\n\t".'"texto": "'.${'texto_'.$i}['texto_'.$i].'",'."\n\t".'"boton-1-color": "'.${"boton_1_Color_".$i}["boton_1_color".$i].'",'."\n\t".'"boton-1-texto": "'.${"boton_1_Texto_".$i}["boton_1_texto".$i].'",'."\n\t".'"url-boton-1": "'.${"url_Boton_1_".$i}["url_Boton_1_".$i].'",'."\n\t".'"boton-2-color": "'.${"boton_2_Color_".$i}["boton_2_color".$i].'",'."\n\t".'"boton-2-texto": "'.${"boton_2_Texto_".$i}["boton_2_texto".$i].'",'."\n\t".'"url-boton-2": "'.${"url_Boton_2_".$i}["url_Boton_2_".$i].'",'."\n\t".'"foto-delante": "'.${"ruta_Foto_Delante_".$i}.'",'."\n\t".'"fondo": "'.${"ruta_Fondo_".$i}.'"'."\n".'},{';
             }
 
-            
+
         }
 
         $actualizar = array('carrusel' => $carrusel);
@@ -250,7 +269,7 @@ class CarruselController extends Controller
 
 
     }
-    
+
     /*=====  End of Agregar nuevo item al carrusel  ======*/
-    
+
 }
